+++ src/ReactionTypes.jl
@@ -15,20 +15,20 @@ export check_eq_reaction
 abstract type AbstractReaction end
 
 # mutable struct NBodyReaction <: AbstractReaction
-#     reactant::Vector{Vector{Int32}}
-#     product::Vector{Vector{Int32}}
+#     reactant::Vector{Vector{Int}}
+#     product::Vector{Vector{Int}}
 # end
 
 mutable struct ProbDecay <: AbstractReaction
-    reactant::Vector{Tuple{Int64, Int64}} # [(Z_0, N_0), (Z_1, N_1), ... (Z_n, N_n)]
-    product::Vector{Tuple{Int64, Int64}}
+    reactant::Vector{Tuple{Int, Int}} # [(Z_0, N_0), (Z_1, N_1), ... (Z_n, N_n)]
+    product::Vector{Tuple{Int, Int}}
     rate::Float64
     average_number::Vector{Float64}
 end
 
 mutable struct NeutronCapture <: AbstractReaction # Temperature Dependent Reactions
-    reactant::Vector{Tuple{Int64, Int64}} # [(Z_0, N_0), (Z_1, N_1), ... (Z_n, N_n)]
-    product::Vector{Tuple{Int64, Int64}}
+    reactant::Vector{Tuple{Int, Int}} # [(Z_0, N_0), (Z_1, N_1), ... (Z_n, N_n)]
+    product::Vector{Tuple{Int, Int}}
     rate::Interpolations.Extrapolation
     pfunc::Interpolations.Extrapolation
     current_rate::Float64 
@@ -36,8 +36,8 @@ mutable struct NeutronCapture <: AbstractReaction # Temperature Dependent Reacti
 end
 
 mutable struct AlphaDecay <: AbstractReaction
-    reactant::Tuple{Int64, Int64} # (Z, N)
-    product::Vector{Tuple{Int64, Int64}} # [(Z_0, N_0), (Z_1, N_1)]
+    reactant::Tuple{Int, Int} # (Z, N)
+    product::Vector{Tuple{Int, Int}} # [(Z_0, N_0), (Z_1, N_1)]
     rate::Float64
 end
 
@@ -52,13 +52,13 @@ end
 
 mutable struct ReactionData
     probdecay::Vector{ProbDecay}
-    neutroncapture::Dict{Int64, NeutronCapture}
+    neutroncapture::Dict{Int, NeutronCapture}
     alphadecay::Vector{AlphaDecay}
 end
 
 function initialize_reactions()
     probdecay = Vector{ProbDecay}()
-    ncap_dict = Dict{Int64, NeutronCapture}()
+    ncap_dict = Dict{Int, NeutronCapture}()
     alphadecay = Vector{AlphaDecay}()
     return ReactionData(probdecay, ncap_dict, alphadecay)
 end
